Moved all activeMQ files over to new plugin
