+++ base/uk.ac.stfc.isis.ibex.activemq/build.properties
@@ -1,4 +1,5 @@
 source.. = src/
 output.. = bin/
 bin.includes = .,\
-               META-INF/
+               META-INF/,\
+               lib/activemq-all-5.4.2.jar
