+++ base/uk.ac.stfc.isis.ibex.activemq/src/uk/ac/stfc/isis/ibex/activemq/message/LogMessageFields.java
@@ -16,7 +16,7 @@
  * http://opensource.org/licenses/eclipse-1.0.php
  */
 
-package uk.ac.stfc.isis.ibex.log.message;
+package uk.ac.stfc.isis.ibex.activemq.message;
 
 /**
  * Names for log message fields as used in the XML representations.
