+++ base/uk.ac.stfc.isis.ibex.log/build.properties
@@ -2,6 +2,5 @@ source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
-               lib/activemq-all-5.4.2.jar,\
                plugin.xml,\
                lib/mysql-connector-java-5.1.31-bin.jar
