+++ base/uk.ac.stfc.isis.ibex.log/src/uk/ac/stfc/isis/ibex/log/LogCounter.java
@@ -20,7 +20,8 @@ package uk.ac.stfc.isis.ibex.log;
 
 import org.eclipse.jface.preference.IPreferenceStore;
 
-import uk.ac.stfc.isis.ibex.log.message.LogMessage;
+import uk.ac.stfc.isis.ibex.activemq.ILogMessageConsumer;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessage;
 import uk.ac.stfc.isis.ibex.log.preferences.PreferenceConstants;
 import uk.ac.stfc.isis.ibex.model.ModelObject;
