+++ base/uk.ac.stfc.isis.ibex.log/src/uk/ac/stfc/isis/ibex/log/rdb/LogMessageQuery.java
@@ -26,10 +26,10 @@ import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessage;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessageFields;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessageFieldsSql;
 import uk.ac.stfc.isis.ibex.databases.Rdb;
-import uk.ac.stfc.isis.ibex.log.message.LogMessage;
-import uk.ac.stfc.isis.ibex.log.message.LogMessageFields;
-import uk.ac.stfc.isis.ibex.log.message.sql.LogMessageFieldsSql;
 import uk.ac.stfc.isis.ibex.log.message.sql.LogMessageFieldsWhereSql;
 import uk.ac.stfc.isis.ibex.log.message.sql.LogMessageSql;
