+++ base/uk.ac.stfc.isis.ibex.ui.dae/META-INF/MANIFEST.MF
@@ -21,7 +21,8 @@ Require-Bundle: org.eclipse.ui,
  uk.ac.stfc.isis.ibex.log;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.ui.log;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.ui.perspectives,
- uk.ac.stfc.isis.ibex.epics
+ uk.ac.stfc.isis.ibex.epics,
+ uk.ac.stfc.isis.ibex.activemq
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: uk.ac.stfc.isis.ibex.ui.dae
