+++ base/uk.ac.stfc.isis.ibex.ui.log/META-INF/MANIFEST.MF
@@ -6,6 +6,7 @@ Bundle-Version: 1.0.0.qualifier
 Bundle-Activator: uk.ac.stfc.isis.ibex.ui.log.Activator
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
+ uk.ac.stfc.isis.ibex.activemq;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.log;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.model;bundle-version="1.0.0",
  com.google.common;bundle-version="10.0.1",
