+++ base/uk.ac.stfc.isis.ibex.ui.log/src/uk/ac/stfc/isis/ibex/ui/log/widgets/ISearchModel.java
@@ -21,7 +21,7 @@ package uk.ac.stfc.isis.ibex.ui.log.widgets;
 
 import java.util.Calendar;
 
-import uk.ac.stfc.isis.ibex.log.message.LogMessageFields;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessageFields;
 
 public interface ISearchModel {
 	void search(LogMessageFields field, String value, Calendar from, Calendar to);
