+++ base/uk.ac.stfc.isis.ibex.ui.log/src/uk/ac/stfc/isis/ibex/ui/log/widgets/LogDisplayModel.java
@@ -27,10 +27,10 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
-import uk.ac.stfc.isis.ibex.log.ILogMessageConsumer;
-import uk.ac.stfc.isis.ibex.log.ILogMessageProducer;
-import uk.ac.stfc.isis.ibex.log.message.LogMessage;
-import uk.ac.stfc.isis.ibex.log.message.LogMessageFields;
+import uk.ac.stfc.isis.ibex.activemq.ILogMessageConsumer;
+import uk.ac.stfc.isis.ibex.activemq.ILogMessageProducer;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessage;
+import uk.ac.stfc.isis.ibex.activemq.message.LogMessageFields;
 import uk.ac.stfc.isis.ibex.model.ModelObject;
 import uk.ac.stfc.isis.ibex.ui.AsyncMessageModeratorTask;
 import uk.ac.stfc.isis.ibex.ui.log.comparator.LogMessageComparator;
